Vueless proposal using RequireJS (option 1)
