+++ .travis.yml
@@ -5,7 +5,6 @@ services:
 script:
   - docker-compose build
   - docker-compose run backend lein test
-  - docker-compose run frontend yarn test
   - docker-compose run testcafe firefox /tests/test.js --app 'true' --app-init-delay 20000
 
 notifications:
